moved CreateAnnouncement.kt in notifications package, and put the method to add announcement in the notificationsViewModel
